requirements.txt
streamlit
pandas
requests
python-dotenv
plotly